Make a separate instance of placeholder values for `Runtime`
