+++ java/src/main/kotlin/io/spine/validation/java/setonce/SetOnceConstraintViolation.kt
@@ -37,15 +37,15 @@ import io.spine.protodata.java.mapExpression
 import io.spine.protodata.java.newBuilder
 import io.spine.protodata.java.packToAny
 import io.spine.validate.ConstraintViolation
-import io.spine.validate.ErrorPlaceholder
-import io.spine.validate.ErrorPlaceholder.FIELD_PATH
-import io.spine.validate.ErrorPlaceholder.FIELD_PROPOSED_VALUE
-import io.spine.validate.ErrorPlaceholder.FIELD_TYPE
-import io.spine.validate.ErrorPlaceholder.FIELD_VALUE
-import io.spine.validate.ErrorPlaceholder.PARENT_TYPE
 import io.spine.validate.TemplateString
 import io.spine.validate.checkPlaceholdersHasValue
 import io.spine.validation.IF_SET_AGAIN
+import io.spine.validation.java.ErrorPlaceholder
+import io.spine.validation.java.ErrorPlaceholder.FIELD_PATH
+import io.spine.validation.java.ErrorPlaceholder.FIELD_PROPOSED_VALUE
+import io.spine.validation.java.ErrorPlaceholder.FIELD_TYPE
+import io.spine.validation.java.ErrorPlaceholder.FIELD_VALUE
+import io.spine.validation.java.ErrorPlaceholder.PARENT_TYPE
 
 /**
  * Builds a [ConstraintViolation] instance for the given field.
